speed it up

+++ RecoEcal/EgammaClusterAlgos/interface/EgammaSCEnergyCorrectionAlgo.h
@@ -23,7 +23,7 @@ public:
   enum VerbosityLevel { pDEBUG = 0, pWARNING = 1, pINFO = 2, pERROR = 3 }; 
   
   // public member functions
-  EgammaSCEnergyCorrectionAlgo(double noise, 
+  EgammaSCEnergyCorrectionAlgo(float noise, 
 			       reco::CaloCluster::AlgoId theAlgo,
 			       const edm::ParameterSet& pset, 
 			       VerbosityLevel verbosity = pERROR
@@ -53,7 +53,7 @@ private:
   // 2sigma noise level
   int nCrystalsGT2Sigma(reco::BasicCluster const & seed, EcalRecHitCollection const & rhc) const;
     
-  double sigmaElectronicNoise_;
+  float sigmaElectronicNoise_;
     
   //  the verbosity level
   VerbosityLevel verbosity_;
